% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeShinyApp.R
\name{initializeShinyApp}
\alias{initializeShinyApp}
\title{Initialize a themed shiny app}
\usage{
initializeShinyApp(dest = "./", theme = "SALURBAL")
}
\arguments{
\item{dest}{the path to initialize your app. Important: you should not specify the file name, just the directory.
'./folder/app.R' is wrong. './folder/' is correct. A folder will be created with an app.R file already initialized!}

\item{theme}{which theme do you want: Two options are available "UHC" or "SALURBAL"}
}
\value{
A character vector.
}
\description{
We are working to make high level function which will initialize a \strong{boilerplate shiny application} with certain
features already included; for now these \strong{include branding and a loading page}. In the future we will work to
add additional modules as specified by the user. For example an the addition of boilerplate code for
sync-maps, sankeys ... any other type of visualization that we have previous already developed. This
hopefully will allow open-sourcing of high level shiny development and making getting started with Shiny easier.
\strong{Right now two modules are loaded: 1) a generic ggplot histogram 2) a sync-map component.}
}
\examples{
initializeShinyApp(dest = "./demo", theme = "SALURBAL")
}
