% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_zcta_philly.R
\docType{data}
\name{sf_zcta_philly}
\alias{sf_zcta_philly}
\title{Boundary files for Philadelphia ZCTAS}
\format{
An sf Object with four columns
\describe{
\item{zcta}{a string with five characters which contain the ZCTA id}
\item{place}{a string with seven characters which contain the place id}
\item{city}{name of city, which is just Philadelphia}
\item{geometry}{geometries of each zcta}
}
}
\source{
{Adélie penguins: Palmer Station Antarctica LTER and K. Gorman. 2020. Structural size measurements and isotopic signatures of foraging among adult male and female Adélie penguins (Pygoscelis adeliae) nesting along the Palmer Archipelago near Palmer Station, 2007-2009 ver 5. Environmental Data Initiative} \url{https://doi.org/10.6073/pasta/98b16d7d563f265cb52372c8ca99e60f}
}
\usage{
sf_zcta_philly
}
\description{
Boundary files for Philadelphia ZCTAS
}
\keyword{datasets}
