% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeShinyApp.R
\name{initializeShinyApp}
\alias{initializeShinyApp}
\title{Initialize a themed shiny app}
\usage{
initializeShinyApp(dest = "./", theme = "SALURBAL")
}
\arguments{
\item{dest}{the path to initialize your app. Important: you should not specify the file name, just the directory.
'./folder/app.R' is wrong. './folder/' is correct. A folder will be created with an app.R file already initialized!}

\item{theme}{which theme do you want: Two options are available "UHC" or "SALURBAL"}
}
\value{
A character vector.
}
\description{
Initialize a themed shiny app
}
\examples{
initializeShinyApp(dest = "./", theme = "SALURBAL")
}
