% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeShinyApp.R
\name{initializeShinyApp}
\alias{initializeShinyApp}
\title{Split a string}
\usage{
initializeShinyApp(dest = "./", theme = "SALURBAL")
}
\arguments{
\item{dest}{the path to initialize your app. By default it is the current directory}

\item{theme}{which theme do you want: 1) "SALURBAL" or "UHC"}
}
\value{
A character vector.
}
\description{
Split a string
}
\examples{
initializeShinyApp(dest = "./", theme = "SALURBAL")
}
